use uint64

+++ src/Common/Common.go
@@ -7,8 +7,8 @@ import (
 	"strings"
 )
 
-func BitNum(num uint64) int32 { //the number of bits needed by the num; 0 needs 0, 1 need 1, 2 need 2, 3 need 2 ....
-	var bitn int32 = 63
+func BitNum(num uint64) uint64 { //the number of bits needed by the num; 0 needs 0, 1 need 1, 2 need 2, 3 need 2 ....
+	var bitn uint64 = 63
 	for (bitn >= 0) && (((uint64(1) << uint32(bitn)) & num) == 0) {
 		bitn--
 	}
